requirements.txt
python-docx
python-dotenv
google-genai
Pillow
dotenv
google.generativeai